Playwright_Study_week1/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
async-timeout==5.0.1
attrs==24.2.0
baidu-aip==4.16.13
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
clickhouse-cli==0.3.9
exceptiongroup==1.2.2
furl==2.1.3
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
jsonpath==0.82.2
logging_tree==1.10
logx==0.1.1579232358
lxml==5.3.0
Markdown==3.7
migra==3.0.1663481299
orderedmultidict==1.0.1
packaging==24.2
pendulum==3.0.0
playwright==1.49.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyee==12.0.0
Pygments==2.18.0
PyMySQL==1.1.1
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.6.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
results==0.1.1588569394
schemainspect==3.1.1663587362
six==1.17.0
SQLAlchemy==2.0.36
sqlbag==0.1.1617247075
sqlparse==0.4.3
tabulate==0.9.0
text-unidecode==1.3
time-machine==2.16.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.21.0
